prompt-eval-sdk/requirements.txt
azure-ai-evaluation
promptflow
promptflow-tools
requests
pytest
pytest-mock
setuptools
